Implement A-Gradle. Partial stylecheck compliance.
