+++ .gitignore
@@ -15,6 +15,7 @@ bin/
 
 # data file for saving tasks
 /data/tasks.txt
+/config/
 
 /text-ui-test/ACTUAL.TXT
 text-ui-test/EXPECTED-UNIX.TXT
