+++ build.gradle
@@ -2,6 +2,10 @@ plugins {
     id 'java'
     id 'application'
     id 'com.github.johnrengelman.shadow' version '7.1.2'
+    id 'checkstyle'
+}
+checkstyle {
+    toolVersion = '10.2'
 }
 
 repositories {
@@ -28,7 +32,7 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("Duke.Duke")
 }
 
 shadowJar {
