+++ src/main/java/command/AddTaskCommand.java
@@ -1,10 +1,11 @@
-package Command;
-import DukeException.FailureInExecuteException;
-import Duke.TaskList;
-import Duke.UserInterface;
-import Task.Task;
+package command;
 
 import java.io.IOException;
+
+import duke.TaskList;
+import duke.UserInterface;
+import dukeexception.FailureInExecuteException;
+import task.Task;
 
 public class AddTaskCommand implements Commandable {
     private Task task;
@@ -12,10 +13,11 @@ public class AddTaskCommand implements Commandable {
     public void setTask(Task task) {
         this.task = task;
     }
+
     public boolean execute(TaskList list, UserInterface ui) throws FailureInExecuteException {
         try {
             list.addTask(task);
-        } catch(IOException e) {
+        } catch (IOException e) {
             throw new FailureInExecuteException(e.getMessage());
         }
         return false;
