+++ src/main/java/command/ListCommand.java
@@ -1,6 +1,6 @@
-package Command;
-import Duke.TaskList;
-import Duke.UserInterface;
+package command;
+import duke.TaskList;
+import duke.UserInterface;
 
 public class ListCommand implements Commandable {
     public boolean execute(TaskList list, UserInterface ui) {
