+++ src/main/java/command/MarkCommand.java
@@ -1,18 +1,23 @@
-package Command;
-import DukeException.FailureInExecuteException;
-import Duke.TaskList;
-import Duke.UserInterface;
+package command;
+
 import java.io.IOException;
 
+import duke.TaskList;
+import duke.UserInterface;
+import dukeexception.FailureInExecuteException;
+
 public class MarkCommand implements Commandable {
-    boolean isToBeMarked;
-    int targetIndex;
+    private boolean isToBeMarked;
+    private int targetIndex;
+
     public MarkCommand(boolean isToBeMarked) {
         this.isToBeMarked = isToBeMarked;
     }
+
     public void setMark(int targetIndex) {
         this.targetIndex = targetIndex;
     }
+
     public boolean execute(TaskList list, UserInterface ui) throws FailureInExecuteException {
         try {
             if (!list.setMark(targetIndex, isToBeMarked)) {
