+++ src/main/java/command/ShutdownCommand.java
@@ -1,6 +1,6 @@
-package Command;
-import Duke.TaskList;
-import Duke.UserInterface;
+package command;
+import duke.TaskList;
+import duke.UserInterface;
 
 public class ShutdownCommand implements Commandable {
     public boolean execute(TaskList list, UserInterface ui) {
