+++ src/main/java/duke/Duke.java
@@ -1,10 +1,12 @@
-package Duke;
-import Command.Commandable;
+package duke;
+
 import java.io.IOException;
-import DukeException.InvalidCommandException;
-import DukeException.CorruptedFileException;
-import DukeException.FailureInExecuteException;
-import DukeException.InvalidVarException;
+
+import command.Commandable;
+import dukeexception.CorruptedFileException;
+import dukeexception.FailureInExecuteException;
+import dukeexception.InvalidCommandException;
+import dukeexception.InvalidVarException;
 
 public class Duke {
     private TaskList list;
@@ -12,17 +14,17 @@ public class Duke {
     private Parser parser;
     private UserInterface ui;
     private final String filePath = "./data/tasks.txt";
-    public Duke(){
+    private String logo = " _           _        \n"
+        + "| |    _   _| | _____ \n"
+        + "| |   | | | | |/ / _ \\\n"
+        + "| |___| |_| |   <  __/\n"
+        + "|____/ \\__,_|_|\\_\\___|\n";
+    public Duke() {
         storage = new Storage();
         list = new TaskList(storage);
         parser = new Parser();
         ui = new UserInterface();
     }
-    String logo = " _           _        \n"
-            + "| |    _   _| | _____ \n"
-            + "| |   | | | | |/ / _ \\\n"
-            + "| |___| |_| |   <  __/\n"
-            + "|____/ \\__,_|_|\\_\\___|\n";
     public void startDuke() throws IOException, CorruptedFileException {
         ui.output("Hi, I'm \n" + logo);
         storage.init(filePath);
@@ -33,8 +35,8 @@ public class Duke {
     }
 
     public boolean corruptedFileHandler() {
-        ui.output("File not properly formatted;\n" +
-                "Clear corrupted file Y/N?");
+        ui.output("File not properly formatted;\n"
+                + "Clear corrupted file Y/N?");
         while (true) {
             String input = ui.input();
             ui.output(input);
@@ -72,11 +74,9 @@ public class Duke {
             try {
                 Commandable command = luke.parser.parse(input);
                 isShuttingDown = command.execute(luke.list, luke.ui);
-            }
-            catch (InvalidCommandException e) {
+            } catch (InvalidCommandException e) {
                 luke.ui.output("Unknown command given; " + e.getMessage());
-            }
-            catch (InvalidVarException e) {
+            } catch (InvalidVarException e) {
                 luke.ui.output("Invalid input; " + e.getMessage());
             } catch (FailureInExecuteException e) {
                 luke.ui.output("Failure to execute command; " + e.getMessage());
