+++ src/main/java/dukeexception/DukeException.java
@@ -1,4 +1,5 @@
-package DukeException;
+package dukeexception;
+
 public class DukeException extends Exception {
     public DukeException() {
         super();
