+++ src/main/java/dukeexception/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package DukeException;
+package dukeexception;
 
 public class InvalidCommandException extends DukeException {
     public InvalidCommandException() {
