+++ src/main/java/task/Deadline.java
@@ -1,4 +1,4 @@
-package Task;
+package task;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
@@ -12,10 +12,10 @@ public class Deadline extends Task {
         super(name, completeStatus);
         this.deadline = deadline;
     }
-    public String toString(){
+    public String toString() {
         return "[D]" + super.toString() + " (by: " + deadline.format(DateTimeFormatter.ofPattern("MMM dd yyyy")) + ")";
     }
     public String fileFormat() {
-        return "DL" + divider + super.fileFormat() + divider + deadline;
+        return "DL" + DIVIDER + super.fileFormat() + DIVIDER + deadline;
     }
 }
